move search to own folder
