+++ Integreat/Integreat.Shared/ApplicationObjects/IntegreatModule.cs
@@ -12,6 +12,7 @@ using Integreat.Shared.Pages.Main;
 using Integreat.Shared.Pages.Settings;
 using Integreat.Shared.ViewModels.Events;
 using Integreat.Shared.ViewModels.General;
+using Integreat.Shared.ViewModels.Search;
 using Integreat.Shared.ViewModels.Settings;
 using Integreat.Utilities;
 using ModernHttpClient;
