+++ Integreat/Integreat.Shared/Pages/MainContentPage.xaml
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8" ?>
+﻿<?xml version="1.0" encoding="utf-8" ?>
 <pages:BaseContentPage
     x:Class="Integreat.Shared.Pages.MainContentPage"
     xmlns="http://xamarin.com/schemas/2014/forms"
