+++ Integreat/Integreat.Shared/Pages/Search/SearchListPage.xaml
@@ -1,0 +1,30 @@
+﻿<?xml version="1.0" encoding="UTF-8" ?>
+<pages:BaseContentPage x:Class="Integreat.Shared.Pages.SearchListPage"
+                       xmlns="http://xamarin.com/schemas/2014/forms"
+                       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                       xmlns:converters="clr-namespace:Integreat.Shared.Converters"
+                       xmlns:pages="clr-namespace:Integreat.Shared.Pages.Search"
+                       Title="{Binding Title}" BackgroundColor="{StaticResource backgroundColor}">
+    <ContentPage.Resources>
+        <ResourceDictionary>
+            <converters:UriImageSourceConverter x:Key="ImageConverter" />
+        </ResourceDictionary>
+    </ContentPage.Resources>
+    <ContentPage.Content>
+        <StackLayout>
+            <SearchBar Placeholder="{Binding Title}" Text="{Binding SearchText}" HeightRequest="60"/>
+            <ListView ItemsSource="{Binding FoundPages}" RowHeight="60" HasUnevenRows="False"
+                      Margin="8,0" IsRefreshing="{Binding IsBusy}">
+                <ListView.ItemTemplate>
+                    <DataTemplate>
+                        <ImageCell Command="{Binding OnTapCommand}"
+                                  CommandParameter="{Binding}"
+                                  Detail="{Binding Description}"
+                                   ImageSource="{Binding Thumbnail}"
+                                  Text="{Binding Title}"
+                                  TextColor="{StaticResource textColor}" />
+                    </DataTemplate>
+                </ListView.ItemTemplate>
+            </ListView></StackLayout>
+    </ContentPage.Content>
+</pages:BaseContentPage>
